requirements.txt
psycopg2==2.9.10
django-environ==0.12.0
Django==4.2.22
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
uuid-utils==0.9.0
voluptuous==0.13.1
black==25.1.0
flake8==7.2.0
celery==5.3.6
redis==5.0.1